Run code quality checks as separate jobs
I didn't find an elegant way of doing it all in one job and
continuing execution in case of failure of a step, yet
marking the entire job as failed if at least one step did so.

The current approach is somewhat wasteful, it always spins up
a new container and checks out the code. Think of a solution.